+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## [Unreleased]
 
+- [Compiler Option]()
+    - Remove **--ccol** option
+        - Now (C, Assembly, Objects) files can be passed as positional arguments
+    - Add **--framework** option
+    - Add **--check_only** option
+        - Equivalent of cc ```-fsyntax-only``` option
+    - Fix opaque type mistaken with simple identifier type
+
 - [Opaque type](https://github.com/EruEri/kosu-lang/pull/72)
     - Add pointer to unknown size type (to interop more easily with c library)
     - Fix: exception raised on unknwon type in external declaration